build: fix build errors occurred in the daily-on-main workflow
https://github.com/mirakc/mirakc/actions/runs/9002637562 failed.

`utoipa-swagger-ui` crates depends on `rustls-tls` since
juhaku/utoipa@859473503b036f1fc200e4d8c37cb7bf9ccb8ad9.  This crates uses
`reqwest` in `build.rs` in order to download a swagger-ui zip file.

This change added a new build dependency on `ring` to `mirakc-core` and caused
the failure.

The root cause of the failure was wrong usage of the `CC` environment variable
in `//docker/build-scripts/mirakc.sh`.  `ring` was built for
`x86_64-unknown-linux-gnu`, but we specified `CC` without the target suffix.  As
a result, `armv7l-linux-musleabihf-gcc` was used.

`CC` and `CFLAGS` environment variables should be defined with the target
suffix.